chore: refactor body to header and footer with body
